+++ crates/miden-objects/src/testing/mod.rs
@@ -4,7 +4,7 @@ pub mod account_component;
 pub mod account_id;
 pub mod asset;
 pub mod block;
-pub mod chain_mmr;
 pub mod constants;
 pub mod note;
+pub mod partial_blockchain;
 pub mod storage;