+++ crates/miden-tx/src/testing/mock_chain/mod.rs
@@ -18,18 +18,17 @@ use miden_objects::{
     batch::{ProposedBatch, ProvenBatch},
     block::{
         AccountTree, AccountWitness, BlockAccountUpdate, BlockHeader, BlockInputs, BlockNoteIndex,
-        BlockNoteTree, BlockNumber, NullifierWitness, OutputNoteBatch, ProposedBlock, ProvenBlock,
+        BlockNoteTree, BlockNumber, Blockchain, NullifierTree, NullifierWitness, OutputNoteBatch,
+        ProposedBlock, ProvenBlock,
     },
-    crypto::{
-        dsa::rpo_falcon512::SecretKey,
-        merkle::{Mmr, Smt, SmtProof},
-    },
+    crypto::{dsa::rpo_falcon512::SecretKey, merkle::SmtProof},
     note::{Note, NoteHeader, NoteId, NoteInclusionProof, NoteType, Nullifier},
     testing::account_code::DEFAULT_AUTH_SCRIPT,
     transaction::{
-        ChainMmr, ExecutedTransaction, ForeignAccountInputs, InputNote, InputNotes,
-        OrderedTransactionHeaders, OutputNote, ProvenTransaction, ToInputNoteCommitments,
-        TransactionHeader, TransactionId, TransactionInputs, TransactionScript,
+        ExecutedTransaction, ForeignAccountInputs, InputNote, InputNotes,
+        OrderedTransactionHeaders, OutputNote, PartialBlockchain, ProvenTransaction,
+        ToInputNoteCommitments, TransactionHeader, TransactionId, TransactionInputs,
+        TransactionScript,
     },
 };
 use rand::{Rng, SeedableRng};
@@ -249,13 +248,13 @@ impl PendingObjects {
 #[derive(Debug, Clone)]
 pub struct MockChain {
     /// An append-only structure used to represent the history of blocks produced for this chain.
-    chain: Mmr,
+    chain: Blockchain,
 
     /// History of produced blocks.
     blocks: Vec<ProvenBlock>,
 
     /// Tree containing the latest `Nullifier`'s tree.
-    nullifiers: Smt,
+    nullifiers: NullifierTree,
 
     /// Tree containing the latest state commitment of each account.
     account_tree: AccountTree,
@@ -282,9 +281,9 @@ pub struct MockChain {
 impl Default for MockChain {
     fn default() -> Self {
         MockChain {
-            chain: Mmr::default(),
+            chain: Blockchain::default(),
             blocks: vec![],
-            nullifiers: Smt::default(),
+            nullifiers: NullifierTree::default(),
             account_tree: AccountTree::new(),
             pending_objects: PendingObjects::new(),
             available_notes: BTreeMap::new(),
@@ -441,7 +440,7 @@ impl MockChain {
     {
         let transactions: Vec<_> = txs.into_iter().map(alloc::sync::Arc::new).collect();
 
-        let (batch_reference_block, chain_mmr) =
+        let (batch_reference_block, partial_blockchain) =
             self.get_batch_inputs(transactions.iter().map(|tx| tx.ref_block_num()));
 
         // TODO: Get the actual proofs as part of get_batch_inputs.
@@ -450,7 +449,7 @@ impl MockChain {
         ProposedBatch::new(
             transactions,
             batch_reference_block,
-            chain_mmr,
+            partial_blockchain,
             unauthenticated_note_proofs,
         )
     }
@@ -462,7 +461,7 @@ impl MockChain {
         let (
             transactions,
             block_header,
-            _chain_mmr,
+            _partial_blockchain,
             _unauthenticated_note_proofs,
             id,
             account_updates,
@@ -736,7 +735,7 @@ impl MockChain {
         }
 
         let block_headers = block_headers_map.values().cloned();
-        let mmr = ChainMmr::from_mmr(&self.chain, block_headers).unwrap();
+        let mmr = PartialBlockchain::from_blockchain(&self.chain, block_headers).unwrap();
 
         TransactionInputs::new(
             account,
@@ -753,11 +752,11 @@ impl MockChain {
     pub fn get_batch_inputs(
         &self,
         tx_reference_blocks: impl IntoIterator<Item = BlockNumber>,
-    ) -> (BlockHeader, ChainMmr) {
-        let (batch_reference_block, chain_mmr) =
-            self.latest_selective_chain_mmr(tx_reference_blocks);
+    ) -> (BlockHeader, PartialBlockchain) {
+        let (batch_reference_block, partial_blockchain) =
+            self.latest_selective_partial_blockchain(tx_reference_blocks);
 
-        (batch_reference_block, chain_mmr)
+        (batch_reference_block, partial_blockchain)
     }
 
     /// Gets foreign account inputs to execute FPI transactions.
@@ -800,7 +799,7 @@ impl MockChain {
                 batch.input_notes().iter().filter_map(|note| note.header().map(NoteHeader::id))
             }));
 
-        let (block_reference_block, chain_mmr) = self.latest_selective_chain_mmr(
+        let (block_reference_block, partial_blockchain) = self.latest_selective_partial_blockchain(
             batch_iterator.clone().map(ProvenBatch::reference_block_num).chain(
                 unauthenticated_note_proofs.values().map(|proof| proof.location().block_num()),
             ),
@@ -814,7 +813,7 @@ impl MockChain {
 
         BlockInputs::new(
             block_reference_block,
-            chain_mmr,
+            partial_blockchain,
             account_witnesses,
             nullifier_proofs,
             unauthenticated_note_proofs,
@@ -877,7 +876,8 @@ impl MockChain {
             // TODO: Implement nullifier tree reset once defined at the protocol level.
             for nullifier in self.pending_objects.created_nullifiers.iter() {
                 self.nullifiers
-                    .insert(nullifier.inner(), [current_block_num.into(), ZERO, ZERO, ZERO]);
+                    .mark_spent(*nullifier, BlockNumber::from(current_block_num))
+                    .expect("nullifier should not already be spent");
             }
             let notes_tree = self.pending_objects.build_notes_tree();
 
@@ -974,7 +974,7 @@ impl MockChain {
             }
 
             self.blocks.push(block.clone());
-            self.chain.add(header.commitment());
+            self.chain.push(header.commitment());
             self.reset_pending();
 
             last_block = Some(block);
@@ -991,33 +991,33 @@ impl MockChain {
     // ACCESSORS
     // =========================================================================================
 
-    /// Returns a refernce to the current [`Mmr`] representing the blockchain.
-    pub fn block_chain(&self) -> &Mmr {
+    /// Returns a refernce to the current [`Blockchain`].
+    pub fn block_chain(&self) -> &Blockchain {
         &self.chain
     }
 
-    /// Gets the latest [ChainMmr].
-    pub fn latest_chain_mmr(&self) -> ChainMmr {
-        // We cannot pass the latest block as that would violate the condition in the transaction
-        // inputs that the chain length of the mmr must match the number of the reference block.
+    /// Gets the latest [PartialBlockchain].
+    pub fn latest_partial_blockchain(&self) -> PartialBlockchain {
+        // We have to exclude the latest block because we need to fetch the state of the chain at
+        // that latest block, which does not include itself.
         let block_headers =
             self.blocks.iter().map(|b| b.header()).take(self.blocks.len() - 1).cloned();
 
-        ChainMmr::from_mmr(&self.chain, block_headers).unwrap()
+        PartialBlockchain::from_blockchain(&self.chain, block_headers).unwrap()
     }
 
-    /// Creates a new [`ChainMmr`] with all reference blocks in the given iterator except for the
-    /// latest block header in the chain and returns that latest block header.
+    /// Creates a new [`PartialBlockchain`] with all reference blocks in the given iterator except
+    /// for the latest block header in the chain and returns that latest block header.
     ///
     /// The intended use is for the latest block header to become the reference block of a new
     /// transaction batch or block.
-    pub fn latest_selective_chain_mmr(
+    pub fn latest_selective_partial_blockchain(
         &self,
         reference_blocks: impl IntoIterator<Item = BlockNumber>,
-    ) -> (BlockHeader, ChainMmr) {
+    ) -> (BlockHeader, PartialBlockchain) {
         let latest_block_header = self.latest_block_header().clone();
         // Deduplicate block numbers so each header will be included just once. This is required so
-        // ChainMmr::from_mmr does not panic.
+        // PartialBlockchain::from_blockchain does not panic.
         let reference_blocks: BTreeSet<_> = reference_blocks.into_iter().collect();
 
         // Include all block headers of the reference blocks except the latest block.
@@ -1027,9 +1027,10 @@ impl MockChain {
             .filter(|block_header| block_header.commitment() != latest_block_header.commitment())
             .collect();
 
-        let chain_mmr = ChainMmr::from_mmr(&self.chain, block_headers).unwrap();
+        let partial_blockchain =
+            PartialBlockchain::from_blockchain(&self.chain, block_headers).unwrap();
 
-        (latest_block_header, chain_mmr)
+        (latest_block_header, partial_blockchain)
     }
 
     /// Returns the witnesses for the provided account IDs of the current account tree.
@@ -1055,8 +1056,8 @@ impl MockChain {
         let mut nullifier_proofs = BTreeMap::new();
 
         for nullifier in nullifiers {
-            let proof = self.nullifiers.open(&nullifier.inner());
-            nullifier_proofs.insert(nullifier, NullifierWitness::new(proof));
+            let witness = self.nullifiers.open(&nullifier);
+            nullifier_proofs.insert(nullifier, witness);
         }
 
         nullifier_proofs
@@ -1087,7 +1088,9 @@ impl MockChain {
 
     /// Returns a reference to the latest [`BlockHeader`].
     pub fn latest_block_header(&self) -> BlockHeader {
-        self.blocks[self.chain.forest() - 1].header().clone()
+        let chain_tip =
+            self.chain.chain_tip().expect("chain should contain at least the genesis block");
+        self.blocks[chain_tip.as_usize()].header().clone()
     }
 
     /// Gets a reference to [BlockHeader] with `block_number`.
@@ -1096,7 +1099,7 @@ impl MockChain {
     }
 
     /// Gets a reference to the nullifier tree.
-    pub fn nullifiers(&self) -> &Smt {
+    pub fn nullifiers(&self) -> &NullifierTree {
         &self.nullifiers
     }