+++ crates/miden-tx/src/tests/kernel_tests/test_prologue.rs
@@ -12,10 +12,9 @@ use miden_lib::{
         TransactionKernel,
         memory::{
             ACCT_DB_ROOT_PTR, ACCT_ID_PTR, BLOCK_COMMITMENT_PTR, BLOCK_METADATA_PTR,
-            BLOCK_NUMBER_IDX, CHAIN_COMMITMENT_PTR, CHAIN_MMR_NUM_LEAVES_PTR, CHAIN_MMR_PEAKS_PTR,
-            INIT_ACCT_COMMITMENT_PTR, INIT_NONCE_PTR, INPUT_NOTE_ARGS_OFFSET,
-            INPUT_NOTE_ASSETS_HASH_OFFSET, INPUT_NOTE_ASSETS_OFFSET, INPUT_NOTE_ID_OFFSET,
-            INPUT_NOTE_INPUTS_COMMITMENT_OFFSET, INPUT_NOTE_METADATA_OFFSET,
+            BLOCK_NUMBER_IDX, CHAIN_COMMITMENT_PTR, INIT_ACCT_COMMITMENT_PTR, INIT_NONCE_PTR,
+            INPUT_NOTE_ARGS_OFFSET, INPUT_NOTE_ASSETS_HASH_OFFSET, INPUT_NOTE_ASSETS_OFFSET,
+            INPUT_NOTE_ID_OFFSET, INPUT_NOTE_INPUTS_COMMITMENT_OFFSET, INPUT_NOTE_METADATA_OFFSET,
             INPUT_NOTE_NULLIFIER_SECTION_PTR, INPUT_NOTE_NUM_ASSETS_OFFSET,
             INPUT_NOTE_SCRIPT_ROOT_OFFSET, INPUT_NOTE_SECTION_PTR, INPUT_NOTE_SERIAL_NUM_OFFSET,
             INPUT_NOTES_COMMITMENT_PTR, MemoryOffset, NATIVE_ACCT_CODE_COMMITMENT_PTR,
@@ -23,6 +22,7 @@ use miden_lib::{
             NATIVE_ACCT_STORAGE_COMMITMENT_PTR, NATIVE_ACCT_STORAGE_SLOTS_SECTION_PTR,
             NATIVE_ACCT_VAULT_ROOT_PTR, NATIVE_NUM_ACCT_PROCEDURES_PTR,
             NATIVE_NUM_ACCT_STORAGE_SLOTS_PTR, NOTE_ROOT_PTR, NULLIFIER_DB_ROOT_PTR,
+            PARTIAL_BLOCKCHAIN_NUM_LEAVES_PTR, PARTIAL_BLOCKCHAIN_PEAKS_PTR,
             PREV_BLOCK_COMMITMENT_PTR, PROOF_COMMITMENT_PTR, PROTOCOL_VERSION_IDX, TIMESTAMP_IDX,
             TX_COMMITMENT_PTR, TX_KERNEL_COMMITMENT_PTR, TX_SCRIPT_ROOT_PTR,
         },
@@ -103,7 +103,7 @@ fn test_transaction_prologue() {
 
     global_input_memory_assertions(process, &tx_context);
     block_data_memory_assertions(process, &tx_context);
-    chain_mmr_memory_assertions(process, &tx_context);
+    partial_blockchain_memory_assertions(process, &tx_context);
     account_data_memory_assertions(process, &tx_context);
     input_notes_memory_assertions(process, &tx_context, &note_args);
 }
@@ -225,23 +225,27 @@ fn block_data_memory_assertions(process: &Process, inputs: &TransactionContext) 
     );
 }
 
-fn chain_mmr_memory_assertions(process: &Process, prepared_tx: &TransactionContext) {
-    // update the chain MMR to point to the block against which this transaction is being executed
-    let mut chain_mmr = prepared_tx.tx_inputs().block_chain().clone();
-    chain_mmr.add_block(prepared_tx.tx_inputs().block_header().clone(), true);
+fn partial_blockchain_memory_assertions(process: &Process, prepared_tx: &TransactionContext) {
+    // update the partial blockchain to point to the block against which this transaction is being
+    // executed
+    let mut partial_blockchain = prepared_tx.tx_inputs().block_chain().clone();
+    partial_blockchain.add_block(prepared_tx.tx_inputs().block_header().clone(), true);
 
     assert_eq!(
-        read_root_mem_word(&process.into(), CHAIN_MMR_NUM_LEAVES_PTR)[0],
-        Felt::new(chain_mmr.chain_length().as_u64()),
-        "The number of leaves should be stored at the CHAIN_MMR_NUM_LEAVES_PTR"
+        read_root_mem_word(&process.into(), PARTIAL_BLOCKCHAIN_NUM_LEAVES_PTR)[0],
+        Felt::new(partial_blockchain.chain_length().as_u64()),
+        "The number of leaves should be stored at the PARTIAL_BLOCKCHAIN_NUM_LEAVES_PTR"
     );
 
-    for (i, peak) in chain_mmr.peaks().peaks().iter().enumerate() {
-        // The peaks should be stored at the CHAIN_MMR_PEAKS_PTR
+    for (i, peak) in partial_blockchain.peaks().peaks().iter().enumerate() {
+        // The peaks should be stored at the PARTIAL_BLOCKCHAIN_PEAKS_PTR
         let i: u32 = i.try_into().expect(
             "Number of peaks is log2(number_of_leaves), this value won't be larger than 2**32",
         );
-        assert_eq!(read_root_mem_word(&process.into(), CHAIN_MMR_PEAKS_PTR + i), Word::from(peak));
+        assert_eq!(
+            read_root_mem_word(&process.into(), PARTIAL_BLOCKCHAIN_PEAKS_PTR + i),
+            Word::from(peak)
+        );
     }
 }